utils/convert_image.py
import base64
from io import BytesIO
from PIL import Image


def to_base64(image: Image) -> str:
    buffered = BytesIO()
    image.save(buffered, format="JPEG")
    return base64.b64encode(buffered.getvalue()).decode("utf-8")
